added some more checks in file processing logic in various scripts

+++ docker-entrypoint.d/10-download-themes.sh
@@ -8,18 +8,23 @@ if [ -r ${THEMES_DOWNLOAD_FILE} ] && [ -d ${THEMES_TARGET_DIR} ] ; then
   echo "Found a file with list of themes to download as: ${THEMES_DOWNLOAD_FILE} . Processing file ..."
 
   THEMES_URL_LIST=$(grep -v '\#' ${THEMES_DOWNLOAD_FILE} | grep 'http' | awk '{print $2}')
-  for THEME_URL in ${THEMES_URL_LIST}; do 
-    echo "Downloading theme-file from URL: ${THEME_URL} ... and saving into ${THEMES_TARGET_DIR}/"
-    URL_BASE_NAME=$(basename ${THEME_URL})
-    curl -sL ${THEME_URL} -o ${THEMES_TARGET_DIR}/${URL_BASE_NAME}
 
-    # Unzip the file (quitely) and overwrite without asking,
-    #  , and remove the zip file, all in a subshell, 
-    #  , so scipt's current dir is not affected. 
-    # pwd
-    ( cd ${THEMES_TARGET_DIR} ; unzip -q -o ${URL_BASE_NAME} ; rm ${URL_BASE_NAME} )
-    # pwd
+  if [ -z ${THEMES_URL_LIST+x} ]; then 
+    echo "The themes file - ${THEMES_DOWNLOAD_FILE} is empty, skipping theme downloads ..."
+  else
+    for THEME_URL in ${THEMES_URL_LIST}; do 
+      echo "Downloading theme-file from URL: ${THEME_URL} ... and saving into ${THEMES_TARGET_DIR}/"
+      URL_BASE_NAME=$(basename ${THEME_URL})
+      curl -sL ${THEME_URL} -o ${THEMES_TARGET_DIR}/${URL_BASE_NAME}
+
+      # Unzip the file (quitely) and overwrite without asking,
+      #  , and remove the zip file, all in a subshell, 
+      #  , so scipt's current dir is not affected. 
+      # pwd
+      ( cd ${THEMES_TARGET_DIR} ; unzip -q -o ${URL_BASE_NAME} ; rm ${URL_BASE_NAME} )
+      # pwd
   done
+  fi
 else
 
   echo "File with list of themes to download was not found under themes/ directory."
