+++ docker-entrypoint.d/30-download-themes-from-git.sh
@@ -9,39 +9,44 @@ if [ -r ${THEMES_GIT_REPOS_FILE} ] && [ "${USE_GIT}" == "1" ] && [ -d ${THEMES_T
   echo "Found a file with list of themes to download from git - as: ${THEMES_GIT_REPOS_FILE} . Processing file ..."
 
   THEMES_GIT_URL_LIST=$(grep -v '\#' ${THEMES_GIT_REPOS_FILE} | egrep "http|https" | awk '{print $2}')
-  for THEME_GIT_URL in ${THEMES_GIT_URL_LIST}; do 
 
-    re='^http.*.git$'
-    # Note: double brackets!
-    if [[ ${THEME_GIT_URL} =~ $re ]]; then
+  if [ -z ${THEMES_GIT_URL_LIST+x} ]; then
+    echo "The themes file - ${THEMES_GIT_REPOS_FILE} is empty, skipping themes download ..."
+  else
 
+    for THEME_GIT_URL in ${THEMES_GIT_URL_LIST}; do 
+
+      re='^http.*.git$'
+      # Note: double brackets!
+      if [[ ${THEME_GIT_URL} =~ $re ]]; then
         echo "Syntax of  GIT repository URL ${THEME_GIT_URL} seem to be OK."
       else
         echo "Syntax of GIT repository URL ${THEME_GIT_URL} is not OK." 
-	echo "The URL in the ${THEME_GIT_REPOS_FILE} file needs to be a git repo (so we can clone it)"
+  	echo "The URL in the ${THEME_GIT_REPOS_FILE} file needs to be a git repo (so we can clone it)"
 	echo "Ignoring ${THEME_GIT_URL} ..."
 	continue
-    fi
+      fi
 
-    echo "Cloning repository from GIT URL: ${THEME_GIT_URL} ... and saving into ${THEMES_TARGET_DIR}/"
-    URL_LHS=$(echo ${THEME_GIT_URL} | awk -F '://' '{print $1}')
-    URL_RHS=$(echo ${THEME_GIT_URL} | awk -F '://' '{print $2}')
+      echo "Cloning repository from GIT URL: ${THEME_GIT_URL} ... and saving into ${THEMES_TARGET_DIR}/"
+      URL_LHS=$(echo ${THEME_GIT_URL} | awk -F '://' '{print $1}')
+      URL_RHS=$(echo ${THEME_GIT_URL} | awk -F '://' '{print $2}')
 
 
-    FULL_GIT_REPO_URL=$(echo "${URL_LHS}://${GITHUB_CREDENTIALS}@${URL_RHS}")
-    # echo "Debug - New URL: ${URL_LHS}://${GITHUB_CREDENTIALS}@${URL_RHS}"
-    # echo "Debug - New URL: ${FULL_GIT_REPO_URL}"
+      FULL_GIT_REPO_URL=$(echo "${URL_LHS}://${GITHUB_CREDENTIALS}@${URL_RHS}")
+      # echo "Debug - New URL: ${URL_LHS}://${GITHUB_CREDENTIALS}@${URL_RHS}"
+      # echo "Debug - New URL: ${FULL_GIT_REPO_URL}"
 
-    # remove the '.git' from basename to find out the name of the directory which will be created after git pull.
-    URL_BASE_NAME=$(basename ${THEME_GIT_URL} | sed 's/\.git$//g')
-    # clone the repo (quitely) and move it to target location, all in a subshell, so scipt's current dir is not affected. 
-    ( 
-      # remove existing directory, else git clone will complain
-      rm -fr ${URL_BASE_NAME}
-      git clone ${FULL_GIT_REPO_URL}
-      mv ${URL_BASE_NAME} ${THEMES_TARGET_DIR}/
-    )
-  done
+      # remove the '.git' from basename to find out the name of the directory which will be created after git pull.
+      URL_BASE_NAME=$(basename ${THEME_GIT_URL} | sed 's/\.git$//g')
+      # clone the repo (quitely) and move it to target location, all in a subshell, so scipt's current dir is not affected. 
+      ( 
+        # remove existing directory, else git clone will complain
+        rm -fr ${URL_BASE_NAME}
+        git clone ${FULL_GIT_REPO_URL}
+        mv ${URL_BASE_NAME} ${THEMES_TARGET_DIR}/
+      )
+    done
+  fi 
 else
   echo "Maybe Git is not setup in this container,"
   echo "the file with list of themes to download was not found under themes/ directory,"
